new_installer.py: move POSIX classes/functions
